Added some new commands

+++ cmds/babyshark.js
@@ -1,6 +1,8 @@
 module.exports.run = async (bot, message, args) => {
     message.channel.send('https://youtu.be/XqZsoesa55w?t=22');
-    message.delete();
+    // If not a DM delete the message
+    if(message.channel.type != "dm")
+        message.delete();
 }
 module.exports.help = {
     name: 'babyshark'
