+++ cmds/ping.js
@@ -1,6 +1,8 @@
 module.exports.run = async (bot, message, args) => {
     message.channel.send('Pong! `' + Math.floor(Math.round(bot.ping)) + 'ms`');
-    message.delete();
+    // If not a DM delete the message
+    if(message.channel.type != "dm")
+        message.delete();
 }
 module.exports.help = {
     name: 'ping'
